Tarea2/arquitectura-1/requirements.txt
litestar[redis]
grpcio
grpcio-tools
pybase62
grpcio
grpcio-tools
uvicorn